Move max-threads/single-threaded to test configuration
