+++ test/CMakeLists.txt
@@ -1,7 +1,6 @@
 include_directories(../third_party/catch)
 include_directories(../third_party/tpce-tool/include)
 include_directories(../third_party/sqlite/include)
-include_directories(sqlite)
 include_directories(include)
 
 # Allows loading tests from registered extensions
