+++ test/helpers/test_config.cpp
@@ -83,6 +83,10 @@ bool TestConfiguration::ParseArgument(const string &arg, idx_t argc, char **argv
 		ParseOption("force_restart", Value(true));
 		return true;
 	}
+	if (arg == "--single-threaded") {
+		ParseOption("max_threads", Value::BIGINT(1));
+		return true;
+	}
 	if (StringUtil::StartsWith(arg, "--memory-leak") || StringUtil::StartsWith(arg, "--test-memory-leak")) {
 		ParseOption("test_memory_leaks", Value(true));
 		return true;
