+++ test/sqlite/CMakeLists.txt
@@ -4,13 +4,8 @@ add_definitions(-DDUCKDB_ROOT_DIRECTORY="${PROJECT_SOURCE_DIR}"
                 -DDUCKDB_BUILD_DIRECTORY="${PROJECT_BINARY_DIR}")
 
 set(SQLITE_TEST_RUNNER_SOURCES
-    result_helper.cpp
-    sqllogic_command.cpp
-    sqllogic_test_runner.cpp
-    sqllogic_parser.cpp
-    test_sqllogictest.cpp
-    sqllogic_test_logger.cpp
-    test_config.cpp)
+    result_helper.cpp sqllogic_command.cpp sqllogic_test_runner.cpp
+    sqllogic_parser.cpp test_sqllogictest.cpp sqllogic_test_logger.cpp)
 
 add_library_unity(test_sqlite OBJECT ${SQLITE_TEST_RUNNER_SOURCES})
 set(ALL_OBJECT_FILES
